Improved language strings (#137)
* Improved language strings

A user saw that in the german translation the naming of the rating
strategies are different to those in the related help. And we thought
that the german translation might be a little better, so I changed the
strings here in the original file to be more speaking. What do you
think about it?

* Adjusted tests to comply with new lang strings