Add a default sorting order: by lastname
There's a case that's not entirely clear to me, where a mixture of the standard sorting behaviour of moodle and table pagination result in a bug, where some students are displayed twice - on different pages - and other students are not shown at all.

Forcing a default sorting order resolves this problem. 

Sorting by lastname is probably in line with expectations, at least at our institution.