Fix failing scenario in completion_condition_vote.feature
The Scenario ‘User completes ratingallocate only if they voted’ failed
as of b894dd2ca3334cb (Set unknown completion state only for automatic
tracking). The culprit is the call to is_enabled in the
save_rating_to_db method. The function is_enabled returns either
COMPLETION_ENABLED/COMPLETION_DISABLED when no argument is passed or
COMPLETION_TRACKING_AUTOMATIC/MANUAL when a course-module is passed to
it. COMPLETION_ENABLED is defined to be 1 which is not equal to
COMPLETION_TRACKING_AUTOMATIC so that the if branch is not executed
and therefore the completion state not updated.